Add surround plugin to nvim
Added surround plugin to nvim and also had to change timeout len. Now
surrounds works without taking an eternity for which-key to pop up, but
the downside is that fugitive and surrounds have only 300 ms to work
before which-key pops up since the time-out finishes. These 2 plugins
for some reason use keybinds differently and don't play nice with
which-key or timeout.